Move chip::Logging::LogV to platform-specific implementation (#4741)
* Split out logging support methods per platform.

- Removed mutually exclusive `#defines`
- Removed STDIO_WITH_TIMESTAMPS since it is never set/used

* Move all logging logic from support to platform (to make logic consistent). Android remains

* Fix darwin logging file

* Restyle fixes

* Fix chip logging include

* Large refactor for logging: move into separate platform as a stand alone, add dependencies for some sample applications

* Update logic to include openthread config, make sure efr32 compiles

* Fix android logging config

* Fix logging dependencies in tests

* Restyle fixes

* Re-remove retain logging

* Slight doxygen update for logv

* Rename logging from zephyr to nrfconnect to match target platform

* Remove remaining stdio logging link from previous checkins

* Add log include directory to nrfconnect CMakeFiles

* Switch linux logs to go to stdout instead of syslog. This is for CIRQUE and may not be sane.

* Add logging include dir to esp32 component.mk file

* Update based on code review comments

* Updated with source_set support

* Restyle fixes

* Code review update: include LogV as platform/logging/LogV.h

* Remove platform/logging/include from various makefiles

* Also fix android include - missed last time because of quotes used for those includes

* Fix cc13x2_26x2 logging dependency

* Moved efr32 and esp32 logging.cpp back into platform directory ... will look into moving platform-specific logging back one by one

* Moved qpg6100 logging back into platform-specific directory

* Restyle fixes

* Move cc13x2_26x2 logging.cpp back to platform directory

* Move k32w logging back to platform

* Move nrfconnect/Zephyr logging back to platform

* fix dependency typo

* Moved darwin logging into platform

* Moved linux logging back to platform

* Added TODO on chip test logging binding to stdio

* Restyle fixes
