+++ src/lib/core/BUILD.gn
@@ -27,13 +27,6 @@ buildconfig_header("chip_buildconfig") {
 
   chip_target_style_unix = chip_target_style == "unix"
   chip_target_style_embedded = chip_target_style == "embedded"
-  chip_logging_style_android = chip_logging_style == "android"
-  chip_logging_style_darwin = chip_logging_style == "darwin"
-  chip_logging_style_external = chip_logging_style == "external"
-  chip_logging_style_stdio = chip_logging_style == "stdio"
-  chip_logging_style_stdio_weak = chip_logging_style == "stdio_weak"
-  chip_logging_style_stdio_with_timestamps =
-      chip_logging_style == "stdio_with_timestamps"
 
   chip_config_memory_management_malloc =
       chip_config_memory_management == "malloc"
@@ -54,12 +47,6 @@ buildconfig_header("chip_buildconfig") {
     "CHIP_CONFIG_ENABLE_ARG_PARSER=${chip_config_enable_arg_parser}",
     "CHIP_TARGET_STYLE_UNIX=${chip_target_style_unix}",
     "CHIP_TARGET_STYLE_EMBEDDED=${chip_target_style_embedded}",
-    "CHIP_LOGGING_STYLE_ANDROID=${chip_logging_style_android}",
-    "CHIP_LOGGING_STYLE_DARWIN=${chip_logging_style_darwin}",
-    "CHIP_LOGGING_STYLE_EXTERNAL=${chip_logging_style_external}",
-    "CHIP_LOGGING_STYLE_STDIO=${chip_logging_style_stdio}",
-    "CHIP_LOGGING_STYLE_STDIO_WEAK=${chip_logging_style_stdio_weak}",
-    "CHIP_LOGGING_STYLE_STDIO_WITH_TIMESTAMPS=${chip_logging_style_stdio_with_timestamps}",
     "CHIP_CONFIG_MEMORY_MGMT_MALLOC=${chip_config_memory_management_malloc}",
     "HAVE_MALLOC=${chip_config_memory_management_malloc}",
     "HAVE_FREE=${chip_config_memory_management_malloc}",
