+++ src/lib/core/CHIPConfig.h
@@ -72,21 +72,7 @@
 #include CHIP_PLATFORM_CONFIG_INCLUDE
 #endif
 
-// clang-format off
-
-
 // Profile-specific Configuration Headers
-/*
-#include "CHIPBDXConfig.h"
-
-#include "CHIPDMConfig.h"
-
-#include "CHIPTimeConfig.h"
-
-#include "CHIPEventLoggingConfig.h"
-
-#include "CHIPRMPConfig.h"
-*/
 /**
  *  @def CHIP_CONFIG_ERROR_TYPE
  *
@@ -97,7 +83,7 @@
 #ifndef CHIP_CONFIG_ERROR_TYPE
 #include <stdint.h>
 
-#define CHIP_CONFIG_ERROR_TYPE                             int32_t
+#define CHIP_CONFIG_ERROR_TYPE int32_t
 #endif // CHIP_CONFIG_ERROR_TYPE
 
 /**
@@ -108,7 +94,7 @@
  *
  */
 #ifndef CHIP_CONFIG_NO_ERROR
-#define CHIP_CONFIG_NO_ERROR                               0
+#define CHIP_CONFIG_NO_ERROR 0
 #endif // CHIP_CONFIG_NO_ERROR
 
 /**
@@ -119,7 +105,7 @@
  *
  */
 #ifndef CHIP_CONFIG_ERROR_MIN
-#define CHIP_CONFIG_ERROR_MIN                              4000
+#define CHIP_CONFIG_ERROR_MIN 4000
 #endif // CHIP_CONFIG_ERROR_MIN
 
 /**
@@ -130,7 +116,7 @@
  *
  */
 #ifndef CHIP_CONFIG_ERROR_MAX
-#define CHIP_CONFIG_ERROR_MAX                              4999
+#define CHIP_CONFIG_ERROR_MAX 4999
 #endif // CHIP_CONFIG_ERROR_MAX
 
 /**
@@ -142,7 +128,7 @@
  *
  */
 #ifndef _CHIP_CONFIG_ERROR
-#define _CHIP_CONFIG_ERROR(e)                              (CHIP_ERROR_MIN + (e))
+#define _CHIP_CONFIG_ERROR(e) (CHIP_ERROR_MIN + (e))
 #endif // _CHIP_CONFIG_ERROR
 
 /**
@@ -156,7 +142,7 @@
  *    #CHIP_CONFIG_USE_MICRO_ECC.
  */
 #ifndef CHIP_CONFIG_USE_OPENSSL_ECC
-#define CHIP_CONFIG_USE_OPENSSL_ECC                        1
+#define CHIP_CONFIG_USE_OPENSSL_ECC 1
 #endif // CHIP_CONFIG_USE_OPENSSL_ECC
 
 /**
@@ -171,7 +157,7 @@
  *
  */
 #ifndef CHIP_CONFIG_USE_MICRO_ECC
-#define CHIP_CONFIG_USE_MICRO_ECC                          0
+#define CHIP_CONFIG_USE_MICRO_ECC 0
 #endif // CHIP_CONFIG_USE_MICRO_ECC
 
 #if CHIP_CONFIG_USE_MICRO_ECC && CHIP_CONFIG_USE_OPENSSL_ECC
@@ -202,7 +188,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP160R1
-#define CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP160R1       0
+#define CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP160R1 0
 #endif // CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP160R1
 
 /**
@@ -214,7 +200,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP192R1
-#define CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP192R1       1
+#define CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP192R1 1
 #endif // CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP192R1
 
 /**
@@ -227,7 +213,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP224R1
-#define CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP224R1       1
+#define CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP224R1 1
 #endif // CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP224R1
 
 /**
@@ -241,7 +227,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP256R1
-#define CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP256R1       1
+#define CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP256R1 1
 #endif // CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP256R1
 
 /**
@@ -299,7 +285,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SUPPORT_PASE_CONFIG0_TEST_ONLY
-#define CHIP_CONFIG_SUPPORT_PASE_CONFIG0_TEST_ONLY         0
+#define CHIP_CONFIG_SUPPORT_PASE_CONFIG0_TEST_ONLY 0
 #endif // CHIP_CONFIG_SUPPORT_PASE_CONFIG0_TEST_ONLY
 
 /**
@@ -311,7 +297,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SUPPORT_PASE_CONFIG1
-#define CHIP_CONFIG_SUPPORT_PASE_CONFIG1                   1
+#define CHIP_CONFIG_SUPPORT_PASE_CONFIG1 1
 #endif // CHIP_CONFIG_SUPPORT_PASE_CONFIG1
 
 /**
@@ -327,7 +313,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SUPPORT_PASE_CONFIG2
-#define CHIP_CONFIG_SUPPORT_PASE_CONFIG2                   0
+#define CHIP_CONFIG_SUPPORT_PASE_CONFIG2 0
 #endif // CHIP_CONFIG_SUPPORT_PASE_CONFIG2
 
 /**
@@ -343,7 +329,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SUPPORT_PASE_CONFIG3
-#define CHIP_CONFIG_SUPPORT_PASE_CONFIG3                   0
+#define CHIP_CONFIG_SUPPORT_PASE_CONFIG3 0
 #endif // CHIP_CONFIG_SUPPORT_PASE_CONFIG3
 
 /**
@@ -360,7 +346,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SUPPORT_PASE_CONFIG4
-#define CHIP_CONFIG_SUPPORT_PASE_CONFIG4                   1
+#define CHIP_CONFIG_SUPPORT_PASE_CONFIG4 1
 #endif // CHIP_CONFIG_SUPPORT_PASE_CONFIG4
 
 /**
@@ -376,7 +362,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SUPPORT_PASE_CONFIG5
-#define CHIP_CONFIG_SUPPORT_PASE_CONFIG5                   0
+#define CHIP_CONFIG_SUPPORT_PASE_CONFIG5 0
 #endif // CHIP_CONFIG_SUPPORT_PASE_CONFIG5
 
 /**
@@ -412,9 +398,9 @@
  */
 #ifndef CHIP_CONFIG_PASE_MESSAGE_PAYLOAD_ALIGNMENT
 #if CHIP_CONFIG_USE_MICRO_ECC
-#define CHIP_CONFIG_PASE_MESSAGE_PAYLOAD_ALIGNMENT         1
+#define CHIP_CONFIG_PASE_MESSAGE_PAYLOAD_ALIGNMENT 1
 #else
-#define CHIP_CONFIG_PASE_MESSAGE_PAYLOAD_ALIGNMENT         0
+#define CHIP_CONFIG_PASE_MESSAGE_PAYLOAD_ALIGNMENT 0
 #endif // CHIP_CONFIG_USE_MICRO_ECC
 #endif // CHIP_CONFIG_PASE_MESSAGE_PAYLOAD_ALIGNMENT
 
@@ -428,7 +414,7 @@
  *
  */
 #ifndef CHIP_CONFIG_PASE_RATE_LIMITER_TIMEOUT
-#define CHIP_CONFIG_PASE_RATE_LIMITER_TIMEOUT              15000
+#define CHIP_CONFIG_PASE_RATE_LIMITER_TIMEOUT 15000
 #endif // CHIP_CONFIG_PASE_RATE_LIMITER_TIMEOUT
 
 /**
@@ -447,7 +433,7 @@
  *
  */
 #ifndef CHIP_CONFIG_PASE_RATE_LIMITER_MAX_ATTEMPTS
-#define CHIP_CONFIG_PASE_RATE_LIMITER_MAX_ATTEMPTS         3
+#define CHIP_CONFIG_PASE_RATE_LIMITER_MAX_ATTEMPTS 3
 #endif // CHIP_CONFIG_PASE_RATE_LIMITER_MAX_ATTEMPTS
 
 /**
@@ -481,7 +467,7 @@
  *
  */
 #ifndef CHIP_CONFIG_MEMORY_MGMT_PLATFORM
-#define CHIP_CONFIG_MEMORY_MGMT_PLATFORM      0
+#define CHIP_CONFIG_MEMORY_MGMT_PLATFORM 0
 #endif // CHIP_CONFIG_MEMORY_MGMT_PLATFORM
 
 /**
@@ -499,7 +485,7 @@
  *
  */
 #ifndef CHIP_CONFIG_MEMORY_MGMT_SIMPLE
-#define CHIP_CONFIG_MEMORY_MGMT_SIMPLE        0
+#define CHIP_CONFIG_MEMORY_MGMT_SIMPLE 0
 #endif // CHIP_CONFIG_MEMORY_MGMT_SIMPLE
 
 /**
@@ -517,7 +503,7 @@
  *
  */
 #ifndef CHIP_CONFIG_MEMORY_MGMT_MALLOC
-#define CHIP_CONFIG_MEMORY_MGMT_MALLOC        1
+#define CHIP_CONFIG_MEMORY_MGMT_MALLOC 1
 #endif // CHIP_CONFIG_MEMORY_MGMT_MALLOC
 
 /**
@@ -525,7 +511,8 @@
  */
 
 #if ((CHIP_CONFIG_MEMORY_MGMT_PLATFORM + CHIP_CONFIG_MEMORY_MGMT_SIMPLE + CHIP_CONFIG_MEMORY_MGMT_MALLOC) != 1)
-#error "Please assert exactly one of CHIP_CONFIG_MEMORY_MGMT_PLATFORM, CHIP_CONFIG_MEMORY_MGMT_SIMPLE, or CHIP_CONFIG_MEMORY_MGMT_MALLOC."
+#error                                                                                                                             \
+    "Please assert exactly one of CHIP_CONFIG_MEMORY_MGMT_PLATFORM, CHIP_CONFIG_MEMORY_MGMT_SIMPLE, or CHIP_CONFIG_MEMORY_MGMT_MALLOC."
 #endif // ((CHIP_CONFIG_MEMORY_MGMT_PLATFORM + CHIP_CONFIG_MEMORY_MGMT_SIMPLE + CHIP_CONFIG_MEMORY_MGMT_MALLOC) != 1)
 
 #if !CHIP_CONFIG_MEMORY_MGMT_MALLOC && CHIP_SYSTEM_CONFIG_USE_BSD_IFADDRS
@@ -547,7 +534,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SIMPLE_ALLOCATOR_USE_SMALL_BUFFERS
-#define CHIP_CONFIG_SIMPLE_ALLOCATOR_USE_SMALL_BUFFERS     0
+#define CHIP_CONFIG_SIMPLE_ALLOCATOR_USE_SMALL_BUFFERS 0
 #endif // CHIP_CONFIG_SIMPLE_ALLOCATOR_USE_SMALL_BUFFERS
 
 /**
@@ -558,7 +545,7 @@
  *    for memory-related checks.
  */
 #ifndef CHIP_CONFIG_MEMORY_DEBUG_CHECKS
-#define CHIP_CONFIG_MEMORY_DEBUG_CHECKS     0
+#define CHIP_CONFIG_MEMORY_DEBUG_CHECKS 0
 #endif // CHIP_CONFIG_MEMORY_DEBUG_CHECKS
 
 /**
@@ -577,7 +564,7 @@
  *
  */
 #ifndef CHIP_CONFIG_MEMORY_DEBUG_DMALLOC
-#define CHIP_CONFIG_MEMORY_DEBUG_DMALLOC     0
+#define CHIP_CONFIG_MEMORY_DEBUG_DMALLOC 0
 #endif // CHIP_CONFIG_MEMORY_DEBUG_DMALLOC
 
 /**
@@ -609,7 +596,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SECURITY_MGR_TIME_ALERTS_DUMMY
-#define CHIP_CONFIG_SECURITY_MGR_TIME_ALERTS_DUMMY         1
+#define CHIP_CONFIG_SECURITY_MGR_TIME_ALERTS_DUMMY 1
 #endif // CHIP_CONFIG_SECURITY_MGR_TIME_ALERTS_DUMMY
 
 /**
@@ -625,7 +612,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SECURITY_MGR_TIME_ALERTS_PLATFORM
-#define CHIP_CONFIG_SECURITY_MGR_TIME_ALERTS_PLATFORM      0
+#define CHIP_CONFIG_SECURITY_MGR_TIME_ALERTS_PLATFORM 0
 #endif // CHIP_CONFIG_SECURITY_MGR_TIME_ALERTS_PLATFORM
 
 /**
@@ -666,7 +653,7 @@
  *
  */
 #ifndef CHIP_CONFIG_RNG_IMPLEMENTATION_PLATFORM
-#define CHIP_CONFIG_RNG_IMPLEMENTATION_PLATFORM            0
+#define CHIP_CONFIG_RNG_IMPLEMENTATION_PLATFORM 0
 #endif // CHIP_CONFIG_RNG_IMPLEMENTATION_PLATFORM
 
 /**
@@ -684,7 +671,7 @@
  *
  */
 #ifndef CHIP_CONFIG_RNG_IMPLEMENTATION_CHIPDRBG
-#define CHIP_CONFIG_RNG_IMPLEMENTATION_CHIPDRBG            0
+#define CHIP_CONFIG_RNG_IMPLEMENTATION_CHIPDRBG 0
 #endif // CHIP_CONFIG_RNG_IMPLEMENTATION_CHIPDRBG
 
 /**
@@ -700,17 +687,19 @@
  *
  */
 #ifndef CHIP_CONFIG_RNG_IMPLEMENTATION_OPENSSL
-#define CHIP_CONFIG_RNG_IMPLEMENTATION_OPENSSL             1
+#define CHIP_CONFIG_RNG_IMPLEMENTATION_OPENSSL 1
 #endif // CHIP_CONFIG_RNG_IMPLEMENTATION_OPENSSL
 
 /**
  *  @}
  */
 
-#if ((CHIP_CONFIG_RNG_IMPLEMENTATION_PLATFORM + CHIP_CONFIG_RNG_IMPLEMENTATION_CHIPDRBG + CHIP_CONFIG_RNG_IMPLEMENTATION_OPENSSL) != 1)
-#error "Please assert exactly one of CHIP_CONFIG_RNG_IMPLEMENTATION_PLATFORM, CHIP_CONFIG_RNG_IMPLEMENTATION_CHIPDRBG, or CHIP_CONFIG_RNG_IMPLEMENTATION_OPENSSL."
-#endif // ((CHIP_CONFIG_RNG_IMPLEMENTATION_PLATFORM + CHIP_CONFIG_RNG_IMPLEMENTATION_CHIPDRBG + CHIP_CONFIG_RNG_IMPLEMENTATION_OPENSSL) != 1)
-
+#if ((CHIP_CONFIG_RNG_IMPLEMENTATION_PLATFORM + CHIP_CONFIG_RNG_IMPLEMENTATION_CHIPDRBG +                                          \
+      CHIP_CONFIG_RNG_IMPLEMENTATION_OPENSSL) != 1)
+#error                                                                                                                             \
+    "Please assert exactly one of CHIP_CONFIG_RNG_IMPLEMENTATION_PLATFORM, CHIP_CONFIG_RNG_IMPLEMENTATION_CHIPDRBG, or CHIP_CONFIG_RNG_IMPLEMENTATION_OPENSSL."
+#endif // ((CHIP_CONFIG_RNG_IMPLEMENTATION_PLATFORM + CHIP_CONFIG_RNG_IMPLEMENTATION_CHIPDRBG +
+       // CHIP_CONFIG_RNG_IMPLEMENTATION_OPENSSL) != 1)
 
 /**
  *  @def CHIP_CONFIG_DEV_RANDOM_DRBG_SEED
@@ -725,7 +714,7 @@
  *
  */
 #ifndef CHIP_CONFIG_DEV_RANDOM_DRBG_SEED
-#define CHIP_CONFIG_DEV_RANDOM_DRBG_SEED              0
+#define CHIP_CONFIG_DEV_RANDOM_DRBG_SEED 0
 #endif // CHIP_CONFIG_DEV_RANDOM_DRBG_SEED
 
 /**
@@ -738,10 +727,8 @@
  *
  */
 #ifndef CHIP_CONFIG_DEV_RANDOM_DEVICE_NAME
-#define CHIP_CONFIG_DEV_RANDOM_DEVICE_NAME                 "/dev/urandom"
+#define CHIP_CONFIG_DEV_RANDOM_DEVICE_NAME "/dev/urandom"
 #endif // CHIP_CONFIG_DEV_RANDOM_DEVICE_NAME
-
-
 
 /**
  *  @name chip AES Block Cipher Algorithm Implementation Configuration.
@@ -772,7 +759,7 @@
  *
  */
 #ifndef CHIP_CONFIG_AES_IMPLEMENTATION_PLATFORM
-#define CHIP_CONFIG_AES_IMPLEMENTATION_PLATFORM            0
+#define CHIP_CONFIG_AES_IMPLEMENTATION_PLATFORM 0
 #endif // CHIP_CONFIG_AES_IMPLEMENTATION_PLATFORM
 
 /**
@@ -787,7 +774,7 @@
  *
  */
 #ifndef CHIP_CONFIG_AES_IMPLEMENTATION_OPENSSL
-#define CHIP_CONFIG_AES_IMPLEMENTATION_OPENSSL             1
+#define CHIP_CONFIG_AES_IMPLEMENTATION_OPENSSL 1
 #endif // CHIP_CONFIG_AES_IMPLEMENTATION_OPENSSL
 
 /**
@@ -802,7 +789,7 @@
  *
  */
 #ifndef CHIP_CONFIG_AES_IMPLEMENTATION_AESNI
-#define CHIP_CONFIG_AES_IMPLEMENTATION_AESNI              0
+#define CHIP_CONFIG_AES_IMPLEMENTATION_AESNI 0
 #endif // CHIP_CONFIG_AES_IMPLEMENTATION_AESNI
 
 /**
@@ -817,16 +804,14 @@
  *
  */
 #ifndef CHIP_CONFIG_AES_IMPLEMENTATION_MBEDTLS
-#define CHIP_CONFIG_AES_IMPLEMENTATION_MBEDTLS              0
+#define CHIP_CONFIG_AES_IMPLEMENTATION_MBEDTLS 0
 #endif // CHIP_CONFIG_AES_IMPLEMENTATION_MBEDTLS
 
 /**
  *  @}
  */
 
-#if ((CHIP_CONFIG_AES_IMPLEMENTATION_PLATFORM + \
-      CHIP_CONFIG_AES_IMPLEMENTATION_OPENSSL + \
-      CHIP_CONFIG_AES_IMPLEMENTATION_AESNI + \
+#if ((CHIP_CONFIG_AES_IMPLEMENTATION_PLATFORM + CHIP_CONFIG_AES_IMPLEMENTATION_OPENSSL + CHIP_CONFIG_AES_IMPLEMENTATION_AESNI +    \
       CHIP_CONFIG_AES_IMPLEMENTATION_MBEDTLS) != 1)
 #error "Please assert exactly one CHIP_CONFIG_AES_IMPLEMENTATION_... option."
 #endif
@@ -843,9 +828,8 @@
  *
  */
 #ifndef CHIP_CONFIG_AES_USE_EXPANDED_KEY
-#define CHIP_CONFIG_AES_USE_EXPANDED_KEY                   0
+#define CHIP_CONFIG_AES_USE_EXPANDED_KEY 0
 #endif // CHIP_CONFIG_AES_USE_EXPANDED_KEY
-
 
 /**
  *  @name chip SHA1 and SHA256 Hash Algorithms Implementation Configuration.
@@ -877,7 +861,7 @@
  *
  */
 #ifndef CHIP_CONFIG_HASH_IMPLEMENTATION_PLATFORM
-#define CHIP_CONFIG_HASH_IMPLEMENTATION_PLATFORM           0
+#define CHIP_CONFIG_HASH_IMPLEMENTATION_PLATFORM 0
 #endif // CHIP_CONFIG_HASH_IMPLEMENTATION_PLATFORM
 
 /**
@@ -894,7 +878,7 @@
  *
  */
 #ifndef CHIP_CONFIG_HASH_IMPLEMENTATION_MINCRYPT
-#define CHIP_CONFIG_HASH_IMPLEMENTATION_MINCRYPT           0
+#define CHIP_CONFIG_HASH_IMPLEMENTATION_MINCRYPT 0
 #endif // CHIP_CONFIG_HASH_IMPLEMENTATION_MINCRYPT
 
 /**
@@ -909,7 +893,7 @@
  *
  */
 #ifndef CHIP_CONFIG_HASH_IMPLEMENTATION_OPENSSL
-#define CHIP_CONFIG_HASH_IMPLEMENTATION_OPENSSL            1
+#define CHIP_CONFIG_HASH_IMPLEMENTATION_OPENSSL 1
 #endif // CHIP_CONFIG_HASH_IMPLEMENTATION_OPENSSL
 
 /**
@@ -924,20 +908,17 @@
  *
  */
 #ifndef CHIP_CONFIG_HASH_IMPLEMENTATION_MBEDTLS
-#define CHIP_CONFIG_HASH_IMPLEMENTATION_MBEDTLS            0
+#define CHIP_CONFIG_HASH_IMPLEMENTATION_MBEDTLS 0
 #endif // CHIP_CONFIG_HASH_IMPLEMENTATION_MBEDTLS
 
 /**
  *  @}
  */
 
-#if ((CHIP_CONFIG_HASH_IMPLEMENTATION_PLATFORM + \
-      CHIP_CONFIG_HASH_IMPLEMENTATION_MINCRYPT + \
-      CHIP_CONFIG_HASH_IMPLEMENTATION_OPENSSL  + \
-      CHIP_CONFIG_HASH_IMPLEMENTATION_MBEDTLS) != 1)
+#if ((CHIP_CONFIG_HASH_IMPLEMENTATION_PLATFORM + CHIP_CONFIG_HASH_IMPLEMENTATION_MINCRYPT +                                        \
+      CHIP_CONFIG_HASH_IMPLEMENTATION_OPENSSL + CHIP_CONFIG_HASH_IMPLEMENTATION_MBEDTLS) != 1)
 #error "Please assert exactly one CHIP_CONFIG_HASH_IMPLEMENTATION_... option."
 #endif
-
 
 /**
  *  @name chip key export protocol configuration.
@@ -972,7 +953,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SUPPORT_KEY_EXPORT_CONFIG1
-#define CHIP_CONFIG_SUPPORT_KEY_EXPORT_CONFIG1            1
+#define CHIP_CONFIG_SUPPORT_KEY_EXPORT_CONFIG1 1
 #endif // CHIP_CONFIG_SUPPORT_KEY_EXPORT_CONFIG1
 
 /**
@@ -984,13 +965,12 @@
  *
  */
 #ifndef CHIP_CONFIG_SUPPORT_KEY_EXPORT_CONFIG2
-#define CHIP_CONFIG_SUPPORT_KEY_EXPORT_CONFIG2            1
+#define CHIP_CONFIG_SUPPORT_KEY_EXPORT_CONFIG2 1
 #endif // CHIP_CONFIG_SUPPORT_KEY_EXPORT_CONFIG2
 
 /**
  *  @}
  */
-
 
 /**
  *  @def CHIP_CONFIG_ALLOW_NON_STANDARD_ELLIPTIC_CURVES
@@ -1001,7 +981,7 @@
  *
  */
 #ifndef CHIP_CONFIG_ALLOW_NON_STANDARD_ELLIPTIC_CURVES
-#define CHIP_CONFIG_ALLOW_NON_STANDARD_ELLIPTIC_CURVES     0
+#define CHIP_CONFIG_ALLOW_NON_STANDARD_ELLIPTIC_CURVES 0
 #endif // CHIP_CONFIG_ALLOW_NON_STANDARD_ELLIPTIC_CURVES
 
 /**
@@ -1012,7 +992,7 @@
  *
  */
 #ifndef CHIP_CONFIG_MAX_EC_BITS
-#define CHIP_CONFIG_MAX_EC_BITS                            256
+#define CHIP_CONFIG_MAX_EC_BITS 256
 #endif // CHIP_CONFIG_MAX_EC_BITS
 
 /**
@@ -1023,7 +1003,7 @@
  *
  */
 #ifndef CHIP_CONFIG_MAX_RSA_BITS
-#define CHIP_CONFIG_MAX_RSA_BITS                           4096
+#define CHIP_CONFIG_MAX_RSA_BITS 4096
 #endif // CHIP_CONFIG_MAX_RSA_BITS
 
 /**
@@ -1035,7 +1015,7 @@
  *
  */
 #ifndef CHIP_CONFIG_MAX_PEER_NODES
-#define CHIP_CONFIG_MAX_PEER_NODES                         128
+#define CHIP_CONFIG_MAX_PEER_NODES 128
 #endif // CHIP_CONFIG_MAX_PEER_NODES
 
 /**
@@ -1046,7 +1026,7 @@
  *
  */
 #ifndef CHIP_CONFIG_MAX_CONNECTIONS
-#define CHIP_CONFIG_MAX_CONNECTIONS                        INET_CONFIG_NUM_TCP_ENDPOINTS
+#define CHIP_CONFIG_MAX_CONNECTIONS INET_CONFIG_NUM_TCP_ENDPOINTS
 #endif // CHIP_CONFIG_MAX_CONNECTIONS
 
 /**
@@ -1061,7 +1041,7 @@
  *    inbound and outbound connections.
  */
 #ifndef CHIP_CONFIG_MAX_INCOMING_TCP_CONNECTIONS
-#define CHIP_CONFIG_MAX_INCOMING_TCP_CONNECTIONS           (CHIP_CONFIG_MAX_CONNECTIONS * 4 / 5)
+#define CHIP_CONFIG_MAX_INCOMING_TCP_CONNECTIONS (CHIP_CONFIG_MAX_CONNECTIONS * 4 / 5)
 #endif // CHIP_CONFIG_MAX_INCOMING_TCP_CONNECTIONS
 
 /**
@@ -1077,7 +1057,7 @@
  *    #CHIP_CONFIG_MAX_INCOMING_TCP_CONNECTIONS.
  */
 #ifndef CHIP_CONFIG_MAX_INCOMING_TCP_CON_FROM_SINGLE_IP
-#define CHIP_CONFIG_MAX_INCOMING_TCP_CON_FROM_SINGLE_IP    2
+#define CHIP_CONFIG_MAX_INCOMING_TCP_CON_FROM_SINGLE_IP 2
 #endif // CHIP_CONFIG_MAX_INCOMING_TCP_CON_FROM_SINGLE_IP
 
 /**
@@ -1088,7 +1068,7 @@
  *
  */
 #ifndef CHIP_CONFIG_MAX_SESSION_KEYS
-#define CHIP_CONFIG_MAX_SESSION_KEYS                       CHIP_CONFIG_MAX_CONNECTIONS
+#define CHIP_CONFIG_MAX_SESSION_KEYS CHIP_CONFIG_MAX_CONNECTIONS
 #endif // CHIP_CONFIG_MAX_SESSION_KEYS
 
 /**
@@ -1103,7 +1083,7 @@
  *
  */
 #ifndef CHIP_CONFIG_MAX_APPLICATION_EPOCH_KEYS
-#define CHIP_CONFIG_MAX_APPLICATION_EPOCH_KEYS             4
+#define CHIP_CONFIG_MAX_APPLICATION_EPOCH_KEYS 4
 #endif // CHIP_CONFIG_MAX_APPLICATION_EPOCH_KEYS
 
 /**
@@ -1116,7 +1096,7 @@
  *
  */
 #ifndef CHIP_CONFIG_MAX_APPLICATION_GROUPS
-#define CHIP_CONFIG_MAX_APPLICATION_GROUPS                 8
+#define CHIP_CONFIG_MAX_APPLICATION_GROUPS 8
 #endif // CHIP_CONFIG_MAX_APPLICATION_GROUPS
 
 /**
@@ -1128,7 +1108,7 @@
  *
  */
 #ifndef CHIP_CONFIG_USE_APP_GROUP_KEYS_FOR_MSG_ENC
-#define CHIP_CONFIG_USE_APP_GROUP_KEYS_FOR_MSG_ENC         1
+#define CHIP_CONFIG_USE_APP_GROUP_KEYS_FOR_MSG_ENC 1
 #endif // CHIP_CONFIG_USE_APP_GROUP_KEYS_FOR_MSG_ENC
 
 /**
@@ -1153,7 +1133,7 @@
  *
  */
 #ifndef CHIP_CONFIG_MAX_CACHED_MSG_ENC_APP_KEYS
-#define CHIP_CONFIG_MAX_CACHED_MSG_ENC_APP_KEYS            (CHIP_CONFIG_MAX_APPLICATION_GROUPS + 1)
+#define CHIP_CONFIG_MAX_CACHED_MSG_ENC_APP_KEYS (CHIP_CONFIG_MAX_APPLICATION_GROUPS + 1)
 #endif // CHIP_CONFIG_MAX_CACHED_MSG_ENC_APP_KEYS
 
 #if !(CHIP_CONFIG_MAX_CACHED_MSG_ENC_APP_KEYS > 0 && CHIP_CONFIG_MAX_CACHED_MSG_ENC_APP_KEYS < 256)
@@ -1200,7 +1180,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SUPPORT_PASSCODE_CONFIG1_TEST_ONLY
-#define CHIP_CONFIG_SUPPORT_PASSCODE_CONFIG1_TEST_ONLY     0
+#define CHIP_CONFIG_SUPPORT_PASSCODE_CONFIG1_TEST_ONLY 0
 #endif // CHIP_CONFIG_SUPPORT_PASSCODE_CONFIG1_TEST_ONLY
 
 /**
@@ -1214,7 +1194,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SUPPORT_PASSCODE_CONFIG2
-#define CHIP_CONFIG_SUPPORT_PASSCODE_CONFIG2               1
+#define CHIP_CONFIG_SUPPORT_PASSCODE_CONFIG2 1
 #endif // CHIP_CONFIG_SUPPORT_PASSCODE_CONFIG2
 
 /**
@@ -1230,7 +1210,7 @@
  *
  */
 #ifndef CHIP_CONFIG_DEFAULT_SECURITY_SESSION_ESTABLISHMENT_TIMEOUT
-#define CHIP_CONFIG_DEFAULT_SECURITY_SESSION_ESTABLISHMENT_TIMEOUT  30000
+#define CHIP_CONFIG_DEFAULT_SECURITY_SESSION_ESTABLISHMENT_TIMEOUT 30000
 #endif // CHIP_CONFIG_DEFAULT_SECURITY_SESSION_ESTABLISHMENT_TIMEOUT
 
 /**
@@ -1243,7 +1223,7 @@
  *
  */
 #ifndef CHIP_CONFIG_DEFAULT_SECURITY_SESSION_IDLE_TIMEOUT
-#define CHIP_CONFIG_DEFAULT_SECURITY_SESSION_IDLE_TIMEOUT           15000
+#define CHIP_CONFIG_DEFAULT_SECURITY_SESSION_IDLE_TIMEOUT 15000
 #endif // CHIP_CONFIG_DEFAULT_SECURITY_SESSION_IDLE_TIMEOUT
 
 /**
@@ -1255,7 +1235,7 @@
  *
  */
 #ifndef CHIP_CONFIG_NUM_MESSAGE_BUFS
-#define CHIP_CONFIG_NUM_MESSAGE_BUFS                       16
+#define CHIP_CONFIG_NUM_MESSAGE_BUFS 16
 #endif // CHIP_CONFIG_NUM_MESSAGE_BUFS
 
 /**
@@ -1267,7 +1247,7 @@
  *
  */
 #ifndef CHIP_CONFIG_MAX_UNSOLICITED_MESSAGE_HANDLERS
-#define CHIP_CONFIG_MAX_UNSOLICITED_MESSAGE_HANDLERS       32
+#define CHIP_CONFIG_MAX_UNSOLICITED_MESSAGE_HANDLERS 32
 #endif // CHIP_CONFIG_MAX_UNSOLICITED_MESSAGE_HANDLERS
 
 /**
@@ -1278,9 +1258,8 @@
  *
  */
 #ifndef CHIP_CONFIG_MAX_EXCHANGE_CONTEXTS
-#define CHIP_CONFIG_MAX_EXCHANGE_CONTEXTS                  16
+#define CHIP_CONFIG_MAX_EXCHANGE_CONTEXTS 16
 #endif // CHIP_CONFIG_MAX_EXCHANGE_CONTEXTS
-
 
 /**
  *  @def CHIP_CONFIG_CONNECT_IP_ADDRS
@@ -1291,7 +1270,7 @@
  *
  */
 #ifndef CHIP_CONFIG_CONNECT_IP_ADDRS
-#define CHIP_CONFIG_CONNECT_IP_ADDRS                       4
+#define CHIP_CONFIG_CONNECT_IP_ADDRS 4
 #endif // CHIP_CONFIG_CONNECT_IP_ADDRS
 
 /**
@@ -1306,7 +1285,7 @@
  *
  */
 #ifndef CHIP_CONFIG_DEFAULT_UDP_MTU_SIZE
-#define CHIP_CONFIG_DEFAULT_UDP_MTU_SIZE                   1280
+#define CHIP_CONFIG_DEFAULT_UDP_MTU_SIZE 1280
 #endif // CHIP_CONFIG_DEFAULT_UDP_MTU_SIZE
 
 /**
@@ -1340,7 +1319,7 @@
  *
  */
 #ifndef CHIP_HEADER_RESERVE_SIZE
-#define CHIP_HEADER_RESERVE_SIZE                           38
+#define CHIP_HEADER_RESERVE_SIZE 38
 #endif // CHIP_HEADER_RESERVE_SIZE
 
 /**
@@ -1351,7 +1330,7 @@
  *
  */
 #ifndef CHIP_TRAILER_RESERVE_SIZE
-#define CHIP_TRAILER_RESERVE_SIZE                          20
+#define CHIP_TRAILER_RESERVE_SIZE 20
 #endif // CHIP_TRAILER_RESERVE_SIZE
 
 /**
@@ -1362,7 +1341,7 @@
  *
  */
 #ifndef CHIP_PORT
-#define CHIP_PORT                                          11097
+#define CHIP_PORT 11097
 #endif // CHIP_PORT
 
 /**
@@ -1373,7 +1352,7 @@
  *
  */
 #ifndef CHIP_UNSECURED_PORT
-#define CHIP_UNSECURED_PORT                                11096
+#define CHIP_UNSECURED_PORT 11096
 #endif // CHIP_UNSECURED_PORT
 
 /**
@@ -1383,7 +1362,7 @@
  *    Enable use of an ephemeral UDP source port for locally initiated chip exchanges.
  */
 #ifndef CHIP_CONFIG_ENABLE_EPHEMERAL_UDP_PORT
-#define CHIP_CONFIG_ENABLE_EPHEMERAL_UDP_PORT              0
+#define CHIP_CONFIG_ENABLE_EPHEMERAL_UDP_PORT 0
 #endif // CHIP_CONFIG_ENABLE_EPHEMERAL_UDP_PORT
 
 /**
@@ -1397,7 +1376,7 @@
  *    including message encryption. Because of this it SHOULD NEVER BE ENABLED IN PRODUCTION BUILDS.
  */
 #ifndef CHIP_CONFIG_SECURITY_TEST_MODE
-#define CHIP_CONFIG_SECURITY_TEST_MODE                     0
+#define CHIP_CONFIG_SECURITY_TEST_MODE 0
 #endif // CHIP_CONFIG_SECURITY_TEST_MODE
 
 /**
@@ -1407,7 +1386,7 @@
  *    Enable support for resolving hostnames with a DNS resolver.
  */
 #ifndef CHIP_CONFIG_ENABLE_DNS_RESOLVER
-#define CHIP_CONFIG_ENABLE_DNS_RESOLVER                    (INET_CONFIG_ENABLE_DNS_RESOLVER)
+#define CHIP_CONFIG_ENABLE_DNS_RESOLVER (INET_CONFIG_ENABLE_DNS_RESOLVER)
 #endif // CHIP_CONFIG_ENABLE_DNS_RESOLVER
 
 /**
@@ -1422,7 +1401,7 @@
  *    change in the future to \c TRUE in all cases.
  */
 #ifndef CHIP_CONFIG_RESOLVE_IPADDR_LITERAL
-#define CHIP_CONFIG_RESOLVE_IPADDR_LITERAL                 (CHIP_SYSTEM_CONFIG_USE_SOCKETS)
+#define CHIP_CONFIG_RESOLVE_IPADDR_LITERAL (CHIP_SYSTEM_CONFIG_USE_SOCKETS)
 #endif // CHIP_CONFIG_RESOLVE_IPADDR_LITERAL
 
 /**
@@ -1436,7 +1415,7 @@
  *
  */
 #ifndef CHIP_CONFIG_ENABLE_TARGETED_LISTEN
-#define CHIP_CONFIG_ENABLE_TARGETED_LISTEN                 (!CHIP_SYSTEM_CONFIG_USE_LWIP)
+#define CHIP_CONFIG_ENABLE_TARGETED_LISTEN (!CHIP_SYSTEM_CONFIG_USE_LWIP)
 #endif // CHIP_CONFIG_ENABLE_TARGETED_LISTEN
 
 /**
@@ -1450,7 +1429,7 @@
  *
  */
 #ifndef CHIP_CONFIG_ENABLE_UNSECURED_TCP_LISTEN
-#define CHIP_CONFIG_ENABLE_UNSECURED_TCP_LISTEN            0
+#define CHIP_CONFIG_ENABLE_UNSECURED_TCP_LISTEN 0
 #endif // CHIP_CONFIG_ENABLE_UNSECURED_TCP_LISTEN
 
 /**
@@ -1461,7 +1440,7 @@
  *
  */
 #ifndef CHIP_CONFIG_DEBUG_CERT_VALIDATION
-#define CHIP_CONFIG_DEBUG_CERT_VALIDATION                  1
+#define CHIP_CONFIG_DEBUG_CERT_VALIDATION 1
 #endif // CHIP_CONFIG_DEBUG_CERT_VALIDATION
 
 /**
@@ -1472,7 +1451,7 @@
  *
  */
 #ifndef CHIP_CONFIG_OPERATIONAL_DEVICE_CERT_CURVE_ID
-#define CHIP_CONFIG_OPERATIONAL_DEVICE_CERT_CURVE_ID       (chip::Profiles::Security::kChipCurveId_prime256v1)
+#define CHIP_CONFIG_OPERATIONAL_DEVICE_CERT_CURVE_ID (chip::Profiles::Security::kChipCurveId_prime256v1)
 #endif // CHIP_CONFIG_OPERATIONAL_DEVICE_CERT_CURVE_ID
 
 /**
@@ -1492,7 +1471,7 @@
  *
  */
 #ifndef CHIP_CONFIG_OP_DEVICE_CERT_VALID_DATE_NOT_BEFORE
-#define CHIP_CONFIG_OP_DEVICE_CERT_VALID_DATE_NOT_BEFORE   0x1B9C
+#define CHIP_CONFIG_OP_DEVICE_CERT_VALID_DATE_NOT_BEFORE 0x1B9C
 #endif // CHIP_CONFIG_OP_DEVICE_CERT_VALID_DATE_NOT_BEFORE
 
 /**
@@ -1512,7 +1491,7 @@
  *
  */
 #ifndef CHIP_CONFIG_OP_DEVICE_CERT_VALID_DATE_NOT_AFTER
-#define CHIP_CONFIG_OP_DEVICE_CERT_VALID_DATE_NOT_AFTER    0x6444
+#define CHIP_CONFIG_OP_DEVICE_CERT_VALID_DATE_NOT_AFTER 0x6444
 #endif // CHIP_CONFIG_OP_DEVICE_CERT_VALID_DATE_NOT_AFTER
 
 /**
@@ -1523,7 +1502,7 @@
  *
  */
 #ifndef CHIP_CONFIG_ENABLE_PASE_INITIATOR
-#define CHIP_CONFIG_ENABLE_PASE_INITIATOR                  1
+#define CHIP_CONFIG_ENABLE_PASE_INITIATOR 1
 #endif // CHIP_CONFIG_ENABLE_PASE_INITIATOR
 
 /**
@@ -1535,7 +1514,7 @@
  *
  */
 #ifndef CHIP_CONFIG_ENABLE_PASE_RESPONDER
-#define CHIP_CONFIG_ENABLE_PASE_RESPONDER                  1
+#define CHIP_CONFIG_ENABLE_PASE_RESPONDER 1
 #endif // CHIP_CONFIG_ENABLE_PASE_RESPONDER
 
 /**
@@ -1546,7 +1525,7 @@
  *
  */
 #ifndef CHIP_CONFIG_ENABLE_CASE_INITIATOR
-#define CHIP_CONFIG_ENABLE_CASE_INITIATOR                  1
+#define CHIP_CONFIG_ENABLE_CASE_INITIATOR 1
 #endif // CHIP_CONFIG_ENABLE_CASE_INITIATOR
 
 /**
@@ -1557,7 +1536,7 @@
  *
  */
 #ifndef CHIP_CONFIG_ENABLE_CASE_RESPONDER
-#define CHIP_CONFIG_ENABLE_CASE_RESPONDER                  1
+#define CHIP_CONFIG_ENABLE_CASE_RESPONDER 1
 #endif // CHIP_CONFIG_ENABLE_CASE_RESPONDER
 
 /**
@@ -1570,7 +1549,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SUPPORT_CASE_CONFIG1
-#define CHIP_CONFIG_SUPPORT_CASE_CONFIG1                   1
+#define CHIP_CONFIG_SUPPORT_CASE_CONFIG1 1
 #endif // CHIP_CONFIG_SUPPORT_CASE_CONFIG1
 
 /**
@@ -1582,13 +1561,13 @@
  */
 #ifndef CHIP_CONFIG_DEFAULT_CASE_CURVE_ID
 #if CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP224R1
-#define CHIP_CONFIG_DEFAULT_CASE_CURVE_ID                  (chip::Profiles::Security::kChipCurveId_secp224r1)
+#define CHIP_CONFIG_DEFAULT_CASE_CURVE_ID (chip::Profiles::Security::kChipCurveId_secp224r1)
 #elif CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP256R1
-#define CHIP_CONFIG_DEFAULT_CASE_CURVE_ID                  (chip::Profiles::Security::kChipCurveId_prime256v1)
+#define CHIP_CONFIG_DEFAULT_CASE_CURVE_ID (chip::Profiles::Security::kChipCurveId_prime256v1)
 #elif CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP192R1
-#define CHIP_CONFIG_DEFAULT_CASE_CURVE_ID                  (chip::Profiles::Security::kChipCurveId_prime192v1)
+#define CHIP_CONFIG_DEFAULT_CASE_CURVE_ID (chip::Profiles::Security::kChipCurveId_prime192v1)
 #else
-#define CHIP_CONFIG_DEFAULT_CASE_CURVE_ID                  (chip::Profiles::Security::kChipCurveId_secp160r1)
+#define CHIP_CONFIG_DEFAULT_CASE_CURVE_ID (chip::Profiles::Security::kChipCurveId_secp160r1)
 #endif
 #endif // CHIP_CONFIG_DEFAULT_CASE_CURVE_ID
 
@@ -1596,14 +1575,16 @@
  *  @def CHIP_CONFIG_DEFAULT_CASE_ALLOWED_CURVES
  *
  *  @brief
- *    Default set of ECDH curves allowed to be used in a CASE session (initiating or responding), if not overridden by the application.
+ *    Default set of ECDH curves allowed to be used in a CASE session (initiating or responding), if not overridden by the
+ * application.
  *
  */
 #ifndef CHIP_CONFIG_DEFAULT_CASE_ALLOWED_CURVES
 #if CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP224R1 || CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP256R1
-#define CHIP_CONFIG_DEFAULT_CASE_ALLOWED_CURVES            (chip::Profiles::Security::kChipCurveSet_secp224r1|chip::Profiles::Security::kChipCurveSet_prime256v1)
+#define CHIP_CONFIG_DEFAULT_CASE_ALLOWED_CURVES                                                                                    \
+    (chip::Profiles::Security::kChipCurveSet_secp224r1 | chip::Profiles::Security::kChipCurveSet_prime256v1)
 #else
-#define CHIP_CONFIG_DEFAULT_CASE_ALLOWED_CURVES            (chip::Profiles::Security::kChipCurveSet_All)
+#define CHIP_CONFIG_DEFAULT_CASE_ALLOWED_CURVES (chip::Profiles::Security::kChipCurveSet_All)
 #endif
 #endif // CHIP_CONFIG_DEFAULT_CASE_ALLOWED_CURVES
 
@@ -1626,7 +1607,7 @@
  *
  */
 #ifndef CHIP_CONFIG_MAX_SHARED_SESSIONS_END_NODES
-#define CHIP_CONFIG_MAX_SHARED_SESSIONS_END_NODES          10
+#define CHIP_CONFIG_MAX_SHARED_SESSIONS_END_NODES 10
 #endif // CHIP_CONFIG_MAX_SHARED_SESSIONS_END_NODES
 
 /**
@@ -1638,7 +1619,7 @@
  *
  */
 #ifndef CHIP_CONFIG_MAX_END_NODES_PER_SHARED_SESSION
-#define CHIP_CONFIG_MAX_END_NODES_PER_SHARED_SESSION       10
+#define CHIP_CONFIG_MAX_END_NODES_PER_SHARED_SESSION 10
 #endif // CHIP_CONFIG_MAX_END_NODES_PER_SHARED_SESSION
 
 /**
@@ -1649,7 +1630,7 @@
  *
  */
 #ifndef CHIP_CONFIG_ENABLE_TAKE_INITIATOR
-#define CHIP_CONFIG_ENABLE_TAKE_INITIATOR                  0
+#define CHIP_CONFIG_ENABLE_TAKE_INITIATOR 0
 #endif // CHIP_CONFIG_ENABLE_TAKE_INITIATOR
 
 /**
@@ -1660,7 +1641,7 @@
  *
  */
 #ifndef CHIP_CONFIG_ENABLE_TAKE_RESPONDER
-#define CHIP_CONFIG_ENABLE_TAKE_RESPONDER                  0
+#define CHIP_CONFIG_ENABLE_TAKE_RESPONDER 0
 #endif // CHIP_CONFIG_ENABLE_TAKE_RESPONDER
 
 /**
@@ -1671,7 +1652,7 @@
  *
  */
 #ifndef CHIP_CONFIG_ENABLE_KEY_EXPORT_INITIATOR
-#define CHIP_CONFIG_ENABLE_KEY_EXPORT_INITIATOR            1
+#define CHIP_CONFIG_ENABLE_KEY_EXPORT_INITIATOR 1
 #endif // CHIP_CONFIG_ENABLE_KEY_EXPORT_INITIATOR
 
 /**
@@ -1682,7 +1663,7 @@
  *
  */
 #ifndef CHIP_CONFIG_ENABLE_KEY_EXPORT_RESPONDER
-#define CHIP_CONFIG_ENABLE_KEY_EXPORT_RESPONDER            1
+#define CHIP_CONFIG_ENABLE_KEY_EXPORT_RESPONDER 1
 #endif // CHIP_CONFIG_ENABLE_KEY_EXPORT_RESPONDER
 
 /**
@@ -1718,7 +1699,7 @@
  *
  */
 #ifndef CHIP_CONFIG_REQUIRE_AUTH
-#define CHIP_CONFIG_REQUIRE_AUTH                           1
+#define CHIP_CONFIG_REQUIRE_AUTH 1
 #endif // CHIP_CONFIG_REQUIRE_AUTH
 
 /**
@@ -1735,7 +1716,7 @@
  *
  */
 #ifndef CHIP_CONFIG_REQUIRE_AUTH_DEVICE_CONTROL
-#define CHIP_CONFIG_REQUIRE_AUTH_DEVICE_CONTROL            CHIP_CONFIG_REQUIRE_AUTH
+#define CHIP_CONFIG_REQUIRE_AUTH_DEVICE_CONTROL CHIP_CONFIG_REQUIRE_AUTH
 #endif // CHIP_CONFIG_REQUIRE_AUTH_DEVICE_CONTROL
 
 /**
@@ -1752,7 +1733,7 @@
  *
  */
 #ifndef CHIP_CONFIG_REQUIRE_AUTH_FABRIC_PROV
-#define CHIP_CONFIG_REQUIRE_AUTH_FABRIC_PROV               CHIP_CONFIG_REQUIRE_AUTH
+#define CHIP_CONFIG_REQUIRE_AUTH_FABRIC_PROV CHIP_CONFIG_REQUIRE_AUTH
 #endif // CHIP_CONFIG_REQUIRE_AUTH_FABRIC_PROV
 
 /**
@@ -1769,7 +1750,7 @@
  *
  */
 #ifndef CHIP_CONFIG_REQUIRE_AUTH_NETWORK_PROV
-#define CHIP_CONFIG_REQUIRE_AUTH_NETWORK_PROV              CHIP_CONFIG_REQUIRE_AUTH
+#define CHIP_CONFIG_REQUIRE_AUTH_NETWORK_PROV CHIP_CONFIG_REQUIRE_AUTH
 #endif // CHIP_CONFIG_REQUIRE_AUTH_NETWORK_PROV
 
 /**
@@ -1786,7 +1767,7 @@
  *
  */
 #ifndef CHIP_CONFIG_REQUIRE_AUTH_SERVICE_PROV
-#define CHIP_CONFIG_REQUIRE_AUTH_SERVICE_PROV              CHIP_CONFIG_REQUIRE_AUTH
+#define CHIP_CONFIG_REQUIRE_AUTH_SERVICE_PROV CHIP_CONFIG_REQUIRE_AUTH
 #endif // CHIP_CONFIG_REQUIRE_AUTH_SERVICE_PROV
 
 /**
@@ -1807,7 +1788,7 @@
  *
  */
 #ifndef CHIP_CONFIG_ENABLE_PROVISIONING_BUNDLE_SUPPORT
-#define CHIP_CONFIG_ENABLE_PROVISIONING_BUNDLE_SUPPORT     1
+#define CHIP_CONFIG_ENABLE_PROVISIONING_BUNDLE_SUPPORT 1
 #endif // CHIP_CONFIG_ENABLE_PROVISIONING_BUNDLE_SUPPORT
 
 /**
@@ -1819,7 +1800,7 @@
  *
  */
 #ifndef CHIP_ERROR_LOGGING
-#define CHIP_ERROR_LOGGING                                 1
+#define CHIP_ERROR_LOGGING 1
 #endif // CHIP_ERROR_LOGGING
 
 /**
@@ -1831,7 +1812,7 @@
  *
  */
 #ifndef CHIP_PROGRESS_LOGGING
-#define CHIP_PROGRESS_LOGGING                              1
+#define CHIP_PROGRESS_LOGGING 1
 #endif // CHIP_PROGRESS_LOGGING
 
 /**
@@ -1843,8 +1824,17 @@
  *
  */
 #ifndef CHIP_DETAIL_LOGGING
-#define CHIP_DETAIL_LOGGING                                1
+#define CHIP_DETAIL_LOGGING 1
 #endif // CHIP_DETAIL_LOGGING
+
+/**
+ * CHIP_CONFIG_LOG_MESSAGE_MAX_SIZE
+ *
+ * The maximum size (in bytes) of a log message
+ */
+#ifndef CHIP_CONFIG_LOG_MESSAGE_MAX_SIZE
+#define CHIP_CONFIG_LOG_MESSAGE_MAX_SIZE 256
+#endif
 
 /**
  *  @def CHIP_CONFIG_ENABLE_FUNCT_ERROR_LOGGING
@@ -1856,7 +1846,6 @@
 #ifndef CHIP_CONFIG_ENABLE_FUNCT_ERROR_LOGGING
 #define CHIP_CONFIG_ENABLE_FUNCT_ERROR_LOGGING 0
 #endif // CHIP_CONFIG_ENABLE_FUNCT_ERROR_LOGGING
-
 
 /**
  *  @def CHIP_CONFIG_ENABLE_CONDITION_LOGGING
@@ -1879,7 +1868,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SERVICE_DIR_CONNECT_TIMEOUT_MSECS
-#define CHIP_CONFIG_SERVICE_DIR_CONNECT_TIMEOUT_MSECS      (10000)
+#define CHIP_CONFIG_SERVICE_DIR_CONNECT_TIMEOUT_MSECS (10000)
 #endif // CHIP_CONFIG_SERVICE_DIR_CONNECT_TIMEOUT_MSECS
 
 /**
@@ -1896,7 +1885,7 @@
  *
  */
 #ifndef CHIP_CONFIG_DEFAULT_INCOMING_CONNECTION_IDLE_TIMEOUT
-#define CHIP_CONFIG_DEFAULT_INCOMING_CONNECTION_IDLE_TIMEOUT        15000
+#define CHIP_CONFIG_DEFAULT_INCOMING_CONNECTION_IDLE_TIMEOUT 15000
 #endif // CHIP_CONFIG_DEFAULT_INCOMING_CONNECTION_IDLE_TIMEOUT
 
 /**
@@ -1910,7 +1899,7 @@
  *
  */
 #ifndef CHIP_CONFIG_MSG_COUNTER_SYNC_RESP_TIMEOUT
-#define CHIP_CONFIG_MSG_COUNTER_SYNC_RESP_TIMEOUT          2000
+#define CHIP_CONFIG_MSG_COUNTER_SYNC_RESP_TIMEOUT 2000
 #endif // CHIP_CONFIG_MSG_COUNTER_SYNC_RESP_TIMEOUT
 
 /**
@@ -1922,7 +1911,7 @@
  *
  */
 #ifndef CHIP_CONFIG_TEST
-#define CHIP_CONFIG_TEST                                   0
+#define CHIP_CONFIG_TEST 0
 #endif // CHIP_CONFIG_TEST
 
 /**
@@ -1934,7 +1923,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SHORT_ERROR_STR
-#define CHIP_CONFIG_SHORT_ERROR_STR                        0
+#define CHIP_CONFIG_SHORT_ERROR_STR 0
 #endif // CHIP_CONFIG_SHORT_ERROR_STR
 
 /**
@@ -1960,9 +1949,9 @@
  */
 #ifndef CHIP_CONFIG_ERROR_STR_SIZE
 #if CHIP_CONFIG_SHORT_ERROR_STR
-#define CHIP_CONFIG_ERROR_STR_SIZE                         (5 + 1 + 10 + 3 + 8 + 1)
+#define CHIP_CONFIG_ERROR_STR_SIZE (5 + 1 + 10 + 3 + 8 + 1)
 #else // CHIP_CONFIG_SHORT_ERROR_STR
-#define CHIP_CONFIG_ERROR_STR_SIZE                         256
+#define CHIP_CONFIG_ERROR_STR_SIZE 256
 #endif // CHIP_CONFIG_SHORT_ERROR_STR
 #endif // CHIP_CONFIG_ERROR_STR_SIZE
 
@@ -1975,7 +1964,7 @@
  *
  */
 #ifndef CHIP_CONFIG_CUSTOM_ERROR_FORMATTER
-#define CHIP_CONFIG_CUSTOM_ERROR_FORMATTER                 0
+#define CHIP_CONFIG_CUSTOM_ERROR_FORMATTER 0
 #endif // CHIP_CONFIG_CUSTOM_ERROR_FORMATTER
 
 /**
@@ -1993,7 +1982,7 @@
  *  and decimal values
  */
 #ifndef CHIP_CONFIG_SHORT_FORM_ERROR_VALUE_FORMAT
-#define CHIP_CONFIG_SHORT_FORM_ERROR_VALUE_FORMAT          "0x%08" PRIX32
+#define CHIP_CONFIG_SHORT_FORM_ERROR_VALUE_FORMAT "0x%08" PRIX32
 #endif // CHIP_CONFIG_SHORT_FORM_ERROR_VALUE_FORMAT
 
 /**
@@ -2006,7 +1995,7 @@
  *
  */
 #ifndef CHIP_CONFIG_BLE_PKT_RESERVED_SIZE
-#define CHIP_CONFIG_BLE_PKT_RESERVED_SIZE                  0
+#define CHIP_CONFIG_BLE_PKT_RESERVED_SIZE 0
 #endif // CHIP_CONFIG_BLE_PKT_RESERVED_SIZE
 
 /**
@@ -2018,7 +2007,7 @@
  *
  */
 #ifndef CHIP_CONFIG_ENABLE_SECURITY_DEBUG_FUNCS
-#define CHIP_CONFIG_ENABLE_SECURITY_DEBUG_FUNCS            1
+#define CHIP_CONFIG_ENABLE_SECURITY_DEBUG_FUNCS 1
 #endif // CHIP_CONFIG_ENABLE_SECURITY_DEBUG_FUNCS
 
 /**
@@ -2045,26 +2034,22 @@
  */
 #ifndef CHIP_CONFIG_IsPlatformErrorNonCritical
 #if CHIP_SYSTEM_CONFIG_USE_LWIP
-#define _CHIP_CONFIG_IsPlatformLwIPErrorNonCritical(CODE)         \
-    ((CODE) == chip::System::MapErrorLwIP(ERR_RTE)         || \
-     (CODE) == chip::System::MapErrorLwIP(ERR_MEM))
+#define _CHIP_CONFIG_IsPlatformLwIPErrorNonCritical(CODE)                                                                          \
+    ((CODE) == chip::System::MapErrorLwIP(ERR_RTE) || (CODE) == chip::System::MapErrorLwIP(ERR_MEM))
 #else // !CHIP_SYSTEM_CONFIG_USE_LWIP
-#define _CHIP_CONFIG_IsPlatformLwIPErrorNonCritical(CODE)  0
+#define _CHIP_CONFIG_IsPlatformLwIPErrorNonCritical(CODE) 0
 #endif // !CHIP_SYSTEM_CONFIG_USE_LWIP
 
 #if CHIP_SYSTEM_CONFIG_USE_SOCKETS
-#define _CHIP_CONFIG_IsPlatformPOSIXErrorNonCritical(CODE)        \
-    ((CODE) == chip::System::MapErrorPOSIX(EHOSTUNREACH)   || \
-     (CODE) == chip::System::MapErrorPOSIX(ENETUNREACH)    || \
-     (CODE) == chip::System::MapErrorPOSIX(EADDRNOTAVAIL)  || \
-     (CODE) == chip::System::MapErrorPOSIX(EPIPE))
+#define _CHIP_CONFIG_IsPlatformPOSIXErrorNonCritical(CODE)                                                                         \
+    ((CODE) == chip::System::MapErrorPOSIX(EHOSTUNREACH) || (CODE) == chip::System::MapErrorPOSIX(ENETUNREACH) ||                  \
+     (CODE) == chip::System::MapErrorPOSIX(EADDRNOTAVAIL) || (CODE) == chip::System::MapErrorPOSIX(EPIPE))
 #else // !CHIP_SYSTEM_CONFIG_USE_SOCKETS
-#define _CHIP_CONFIG_IsPlatformPOSIXErrorNonCritical(CODE)  0
+#define _CHIP_CONFIG_IsPlatformPOSIXErrorNonCritical(CODE) 0
 #endif // !CHIP_SYSTEM_CONFIG_USE_SOCKETS
 
-#define CHIP_CONFIG_IsPlatformErrorNonCritical(CODE)              \
-    (_CHIP_CONFIG_IsPlatformPOSIXErrorNonCritical(CODE)        || \
-     _CHIP_CONFIG_IsPlatformLwIPErrorNonCritical(CODE))
+#define CHIP_CONFIG_IsPlatformErrorNonCritical(CODE)                                                                               \
+    (_CHIP_CONFIG_IsPlatformPOSIXErrorNonCritical(CODE) || _CHIP_CONFIG_IsPlatformLwIPErrorNonCritical(CODE))
 #endif // CHIP_CONFIG_IsPlatformErrorNonCritical
 
 /**
@@ -2077,7 +2062,7 @@
  *
  */
 #ifndef CHIP_CONFIG_WILL_OVERRIDE_PLATFORM_MATH_FUNCS
-#define CHIP_CONFIG_WILL_OVERRIDE_PLATFORM_MATH_FUNCS   0
+#define CHIP_CONFIG_WILL_OVERRIDE_PLATFORM_MATH_FUNCS 0
 #endif // CHIP_CONFIG_WILL_OVERRIDE_PLATFORM_MATH_FUNCS
 
 /**
@@ -2134,7 +2119,7 @@
  *
  */
 #ifndef CHIP_CONFIG_PERSISTED_STORAGE_ENC_MSG_CNTR_ID
-#define CHIP_CONFIG_PERSISTED_STORAGE_ENC_MSG_CNTR_ID      "EncMsgCntr"
+#define CHIP_CONFIG_PERSISTED_STORAGE_ENC_MSG_CNTR_ID "EncMsgCntr"
 #endif // CHIP_CONFIG_PERSISTED_STORAGE_ENC_MSG_CNTR_ID
 
 /**
@@ -2145,7 +2130,7 @@
  *
  */
 #ifndef CHIP_CONFIG_PERSISTED_STORAGE_ENC_MSG_CNTR_EPOCH
-#define CHIP_CONFIG_PERSISTED_STORAGE_ENC_MSG_CNTR_EPOCH   0x1000
+#define CHIP_CONFIG_PERSISTED_STORAGE_ENC_MSG_CNTR_EPOCH 0x1000
 #endif // CHIP_CONFIG_PERSISTED_STORAGE_ENC_MSG_CNTR_EPOCH
 
 /**
@@ -2229,7 +2214,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SUPPORT_LEGACY_ADD_NETWORK_MESSAGE
-#define CHIP_CONFIG_SUPPORT_LEGACY_ADD_NETWORK_MESSAGE     1
+#define CHIP_CONFIG_SUPPORT_LEGACY_ADD_NETWORK_MESSAGE 1
 #endif // CHIP_CONFIG_SUPPORT_LEGACY_ADD_NETWORK_MESSAGE
 
 /**
@@ -2246,7 +2231,7 @@
  *
  */
 #ifndef CHIP_CONFIG_ALWAYS_USE_LEGACY_ADD_NETWORK_MESSAGE
-#define CHIP_CONFIG_ALWAYS_USE_LEGACY_ADD_NETWORK_MESSAGE     0
+#define CHIP_CONFIG_ALWAYS_USE_LEGACY_ADD_NETWORK_MESSAGE 0
 #endif // CHIP_CONFIG_ALWAYS_USE_LEGACY_ADD_NETWORK_MESSAGE
 
 /**
@@ -2257,7 +2242,7 @@
  * chip fabric.
  */
 #ifndef CHIP_CONFIG_ENABLE_IFJ_SERVICE_FABRIC_JOIN
-#define CHIP_CONFIG_ENABLE_IFJ_SERVICE_FABRIC_JOIN         0
+#define CHIP_CONFIG_ENABLE_IFJ_SERVICE_FABRIC_JOIN 0
 #endif // CHIP_CONFIG_ENABLE_IFJ_SERVICE_FABRIC_JOIN
 
 /**
@@ -2268,7 +2253,7 @@
  * device connections across all supported transports.
  */
 #ifndef CHIP_CONFIG_PEER_CONNECTION_POOL_SIZE
-#define CHIP_CONFIG_PEER_CONNECTION_POOL_SIZE                   16
+#define CHIP_CONFIG_PEER_CONNECTION_POOL_SIZE 16
 #endif // CHIP_CONFIG_PEER_CONNECTION_POOL_SIZE
 
 /**
@@ -2278,7 +2263,7 @@
  * expired.
  */
 #ifndef CHIP_PEER_CONNECTION_TIMEOUT_MS
-#define CHIP_PEER_CONNECTION_TIMEOUT_MS      120000
+#define CHIP_PEER_CONNECTION_TIMEOUT_MS 120000
 #endif // CHIP_PEER_CONNECTION_TIMEOUT_MS
 
 /**
@@ -2287,48 +2272,45 @@
  * @brief How frequent are peer connections checked for timeouts.
  */
 #ifndef CHIP_PEER_CONNECTION_TIMEOUT_CHECK_FREQUENCY_MS
-#define CHIP_PEER_CONNECTION_TIMEOUT_CHECK_FREQUENCY_MS      5000
+#define CHIP_PEER_CONNECTION_TIMEOUT_CHECK_FREQUENCY_MS 5000
 #endif // CHIP_PEER_CONNECTION_TIMEOUT_CHECK_FREQUENCY_MS
 
 /**
-   *  @def CHIP_CONFIG_MAX_BINDINGS
-   *
-   *  @brief
-   *    Maximum number of simultaneously active bindings per WeaveExchangeManager
-   *    The new single source TimeSync client takes one binding.
-   *    Every WDM one-way subscription takes one binding. Mutual subscription counts as two one-way subscriptions.
-   *    A reserved slot is needed to take an incoming subscription request.
-   *    For a device with 2 mutual subscriptions, and one single source time sync client, it needs 2 x 2 + 1 = 5 bindings at least.
-   *    At least six is needed if it still wants to take new WDM subscriptions under this load.
-   */
-  #ifndef CHIP_CONFIG_MAX_BINDINGS
-  #define CHIP_CONFIG_MAX_BINDINGS                           6
-  #endif // CHIP_CONFIG_MAX_BINDINGS
+ *  @def CHIP_CONFIG_MAX_BINDINGS
+ *
+ *  @brief
+ *    Maximum number of simultaneously active bindings per WeaveExchangeManager
+ *    The new single source TimeSync client takes one binding.
+ *    Every WDM one-way subscription takes one binding. Mutual subscription counts as two one-way subscriptions.
+ *    A reserved slot is needed to take an incoming subscription request.
+ *    For a device with 2 mutual subscriptions, and one single source time sync client, it needs 2 x 2 + 1 = 5 bindings at least.
+ *    At least six is needed if it still wants to take new WDM subscriptions under this load.
+ */
+#ifndef CHIP_CONFIG_MAX_BINDINGS
+#define CHIP_CONFIG_MAX_BINDINGS 6
+#endif // CHIP_CONFIG_MAX_BINDINGS
 
 /**
-   *  @def CHIP_CONFIG_MAX_DEVICE_ADMINS
-   *
-   *  @brief
-   *    Maximum number of administrators that can provision the device. Each admin
-   *    can provision the device with their unique operational credentials and manage
-   *    their access control lists.
-   */
-  #ifndef CHIP_CONFIG_MAX_DEVICE_ADMINS
-  #define CHIP_CONFIG_MAX_DEVICE_ADMINS                      16
-  #endif // CHIP_CONFIG_MAX_DEVICE_ADMINS
+ *  @def CHIP_CONFIG_MAX_DEVICE_ADMINS
+ *
+ *  @brief
+ *    Maximum number of administrators that can provision the device. Each admin
+ *    can provision the device with their unique operational credentials and manage
+ *    their access control lists.
+ */
+#ifndef CHIP_CONFIG_MAX_DEVICE_ADMINS
+#define CHIP_CONFIG_MAX_DEVICE_ADMINS 16
+#endif // CHIP_CONFIG_MAX_DEVICE_ADMINS
 
-  /**
+/**
  * @def CHIP_NON_PRODUCTION_MARKER
  *
  * @brief Defines the name of a mark symbol whose presence signals that the chip code
  * includes development/testing features that should never be used in production contexts.
  */
 #ifndef CHIP_NON_PRODUCTION_MARKER
-#if (CHIP_CONFIG_SECURITY_TEST_MODE || \
-     CHIP_CONFIG_SUPPORT_PASE_CONFIG0_TEST_ONLY || \
-     CHIP_CONFIG_SUPPORT_PASSCODE_CONFIG1_TEST_ONLY || \
-     (!CHIP_CONFIG_REQUIRE_AUTH) || \
-     CHIP_FUZZING_ENABLED)
+#if (CHIP_CONFIG_SECURITY_TEST_MODE || CHIP_CONFIG_SUPPORT_PASE_CONFIG0_TEST_ONLY ||                                               \
+     CHIP_CONFIG_SUPPORT_PASSCODE_CONFIG1_TEST_ONLY || (!CHIP_CONFIG_REQUIRE_AUTH) || CHIP_FUZZING_ENABLED)
 #define CHIP_NON_PRODUCTION_MARKER WARNING__DO_NOT_SHIP__CONTAINS_NON_PRODUCTION_CHIP_CODE
 #endif
 #endif
@@ -2336,5 +2318,3 @@
 #ifdef CHIP_NON_PRODUCTION_MARKER
 extern const char CHIP_NON_PRODUCTION_MARKER[];
 #endif
-
-// clang-format on
