+++ src/lib/support/BUILD.gn
@@ -36,6 +36,10 @@ action("gen_chip_version") {
   ]
 }
 
+source_set("logging_constants") {
+  sources = [ "logging/Constants.h" ]
+}
+
 source_set("chip_version_header") {
   sources = get_target_outputs(":gen_chip_version")
 
@@ -83,7 +87,6 @@ static_library("support") {
     "UnitTestRegistration.h",
     "logging/CHIPLogging.cpp",
     "logging/CHIPLogging.h",
-    "logging/CHIPLoggingLogV.cpp",
     "verhoeff/Verhoeff.cpp",
     "verhoeff/Verhoeff.h",
     "verhoeff/Verhoeff10.cpp",
@@ -92,11 +95,15 @@ static_library("support") {
     "verhoeff/Verhoeff36.cpp",
   ]
 
+  libs = []
+
   cflags = [ "-Wconversion" ]
 
   public_deps = [
     ":chip_version_header",
+    ":logging_constants",
     "${chip_root}/src/lib/core:chip_config_header",
+    "${chip_root}/src/platform/logging:headers",
     "${nlassert_root}:nlassert",
     "${nlio_root}:nlio",
   ]
@@ -108,6 +115,11 @@ static_library("support") {
     "${chip_root}/src/ble:ble_config_header",
     "${chip_root}/src/inet:inet_config_header",
   ]
+
+  # Android has no 'platform' that provides logging
+  if (current_os == "android") {
+    public_deps += [ "${chip_root}/src/platform/logging:android" ]
+  }
 
   public_configs = [
     "${chip_root}/src:includes",
@@ -129,10 +141,6 @@ static_library("support") {
     public_deps += [ "${nlfaultinjection_root}:nlfaultinjection" ]
   }
 
-  libs = []
-  if (chip_logging_style == "android") {
-    libs += [ "log" ]
-  }
   if (chip_config_memory_debug_dmalloc) {
     libs += [ "dmallocthcxx" ]
   }
