+++ src/lib/support/logging/CHIPLogging.h
@@ -36,6 +36,10 @@
 
 #include <core/CHIPConfig.h>
 
+#include <platform/logging/LogV.h>
+
+#include <support/logging/Constants.h>
+
 #include <stdarg.h>
 #include <stdint.h>
 
@@ -63,129 +67,11 @@
 namespace chip {
 namespace Logging {
 
-/**
- *  @enum LogModule
- *
- *  @brief
- *    Identifies a logical section of code that is a source of log
- *    messages.
- *
- *  @note If you add modules or rearrange this list you must update the
- *        ModuleNames tables in ChipLogging.cpp.
- *
- */
-enum LogModule
-{
-    kLogModule_NotSpecified = 0,
+void LogV(uint8_t module, uint8_t category, const char * msg, va_list args);
+void Log(uint8_t module, uint8_t category, const char * msg, ...);
 
-    kLogModule_Inet,
-    kLogModule_Ble,
-    kLogModule_MessageLayer,
-    kLogModule_SecurityManager,
-    kLogModule_ExchangeManager,
-    kLogModule_TLV,
-    kLogModule_ASN1,
-    kLogModule_Crypto,
-    kLogModule_Controller,
-    kLogModule_Alarm,
-    kLogModule_BDX,
-    kLogModule_DataManagement,
-    kLogModule_DeviceControl,
-    kLogModule_DeviceDescription,
-    kLogModule_Echo,
-    kLogModule_FabricProvisioning,
-    kLogModule_NetworkProvisioning,
-    kLogModule_ServiceDirectory,
-    kLogModule_ServiceProvisioning,
-    kLogModule_SoftwareUpdate,
-    kLogModule_TokenPairing,
-    kLogModule_TimeService,
-    kLogModule_Heartbeat,
-    kLogModule_chipSystemLayer,
-    kLogModule_EventLogging,
-    kLogModule_Support,
-    kLogModule_chipTool,
-    kLogModule_Zcl,
-    kLogModule_Shell,
-    kLogModule_DeviceLayer,
-    kLogModule_SetupPayload,
-    kLogModule_AppServer,
-    kLogModule_Discovery,
-
-    kLogModule_Max
-};
-
-/**
- *  @enum LogCategory
- *
- *  @brief
- *    Identifies a category to which an particular error message
- *    belongs.
- *
- */
-enum LogCategory
-{
-    /*!<
-     *   This log category indicates, when passed to SetLogFilter(),
-     *   that no messages should be logged.
-     *
-     */
-    kLogCategory_None = 0,
-
-    /*!<
-     *   Indicates a category of log message that describes an unexpected
-     *   or severe failure.
-     *
-     *   This log category indicates that a logged message describes
-     *   an unexpected or severe failure in the code.
-     *
-     *   It should be used for things such as out-of-resource errors,
-     *   internal inconsistencies, API misuse, etc. In general, errors
-     *   that are expected to occur as part of normal operation, or
-     *   that are largely determined by external factors (e.g. network
-     *   errors, user/operator induced errors, etc.) should be logged
-     *   as kLogCategory_Progress messages, not as kLogCategory_Error
-     *   messages.
-     *
-     */
-    kLogCategory_Error = 1,
-
-    /*!<
-     *   Indicates a category of log message that describes an event
-     *   that marks the start or end of a major activity, or a major
-     *   change in the state of the overall system.
-     *
-     *   It should be reserved for high-level events. Such messages
-     *   should provide the log reader with a good sense of the
-     *   overall activity of the system at any point in time, while
-     *   being minimally verbose. Where necessary such messages should
-     *   include identifiers or other values that can be used to
-     *   correlate messages involving a common actor or subject
-     *   (e.g. connection ids, request ids, etc.) and/or to identify
-     *   types of actions being taken or handled (e.g.  message types,
-     *   requested resource types, error numbers, etc.).
-     *
-     */
-    kLogCategory_Progress = 2,
-
-    /*!<
-     *   Indicates a category of log message that describes detailed
-     *   information about an event or the state of the system.
-     *
-     *   Such messages can be used to provide ancillary information
-     *   not suitable for the kLogCategory_Error and
-     *   kLogCategory_Progress categories.
-     *
-     */
-    kLogCategory_Detail = 3,
-
-    kLogCategory_Max = kLogCategory_Detail,
-};
-
-extern void LogV(uint8_t module, uint8_t category, const char * msg, va_list args);
-extern void Log(uint8_t module, uint8_t category, const char * msg, ...);
-extern uint8_t GetLogFilter();
-extern void SetLogFilter(uint8_t category);
+uint8_t GetLogFilter();
+void SetLogFilter(uint8_t category);
 
 #ifndef CHIP_ERROR_LOGGING
 #define CHIP_ERROR_LOGGING 1
@@ -272,9 +158,8 @@ static constexpr uint16_t kMaxTrailerLen     = 2;
 static constexpr uint16_t kMaxMessagePadding = (chip::Logging::kMaxPrefixLen + chip::Logging::kMaxModuleNameLen +
                                                 chip::Logging::kMaxSeparatorLen + chip::Logging::kMaxTrailerLen);
 
-extern void GetMessageWithPrefix(char * buf, uint8_t bufSize, uint8_t module, const char * msg);
-extern void GetModuleName(char * buf, uint8_t bufSize, uint8_t module);
-extern void PrintMessagePrefix(uint8_t module);
+void GetMessageWithPrefix(char * buf, uint8_t bufSize, uint8_t module, const char * msg);
+void GetModuleName(char * buf, uint8_t bufSize, uint8_t module);
 
 #else
 
@@ -290,7 +175,7 @@ static inline void GetModuleName(char * buf, uint8_t bufSize, uint8_t module)
 
 #endif // _CHIP_USE_LOGGING
 
-extern bool IsCategoryEnabled(uint8_t CAT);
+bool IsCategoryEnabled(uint8_t category);
 
 /**
  *  @def ChipLogIfFalse(aCondition)
