+++ src/platform/Zephyr/Logging.cpp
@@ -1,37 +1,14 @@
-/*
- *
- *    Copyright (c) 2020 Project CHIP Authors
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
+/* See Project CHIP LICENSE file for licensing information. */
 
-/**
- *    @file
- *          Provides implementations for the CHIP logging functions
- *          on Zephyr platforms.
- */
+#include <platform/logging/LogV.h>
 
-#include <platform/internal/CHIPDeviceLayerInternal.h>
-#include <support/logging/CHIPLogging.h>
+#include <core/CHIPConfig.h>
+#include <support/logging/Constants.h>
 
 #include <kernel.h>
 #include <logging/log.h>
 
 #include <cstdio>
-
-using namespace ::chip;
-using namespace ::chip::DeviceLayer;
-using namespace ::chip::DeviceLayer::Internal;
 
 LOG_MODULE_REGISTER(chip, LOG_LEVEL_DBG);
 
@@ -49,63 +26,53 @@ void __attribute__((weak)) OnLogOutput(void) {}
 } // namespace DeviceLayer
 
 namespace Logging {
+namespace Platform {
 
 /**
  * CHIP log output function.
  */
 
-void LogV(uint8_t module, uint8_t category, const char * msg, va_list v)
+void LogV(const char * module, uint8_t category, const char * msg, va_list v)
 {
-    if (!IsCategoryEnabled(category))
-        return;
+    char formattedMsg[CHIP_CONFIG_LOG_MESSAGE_MAX_SIZE];
 
-    {
-        char formattedMsg[CHIP_DEVICE_CONFIG_LOG_MESSAGE_MAX_SIZE];
-        size_t prefixLen = 0;
+    snprintf(formattedMsg, sizeof(formattedMsg), "%u [%s]", k_uptime_get_32(), module);
 
-        // Max size for "[TAG] {UINT32}"
-        constexpr size_t maxPrefixLen = chip::Logging::kMaxModuleNameLen + 10 + 3;
-        static_assert(sizeof(formattedMsg) > maxPrefixLen);
+    // -2 to ensure at least one byte available for vsnprintf below.
+    formattedMsg[sizeof(formattedMsg) - 2] = 0;
 
-        prefixLen += snprintf(formattedMsg, sizeof(formattedMsg), "%u", k_uptime_get_32());
+    size_t prefixLen = strlen(formattedMsg);
 
-        // Form the log prefix, e.g. "[DL] "
-        formattedMsg[prefixLen++] = '[';
-        GetModuleName(formattedMsg + prefixLen, chip::Logging::kMaxModuleNameLen + 1, module);
-        prefixLen                 = strlen(formattedMsg);
-        formattedMsg[prefixLen++] = ']';
-        formattedMsg[prefixLen++] = ' ';
+    // Append the log message.
+    vsnprintf(formattedMsg + prefixLen, sizeof(formattedMsg) - prefixLen, msg, v);
 
-        // Append the log message.
-        vsnprintf(formattedMsg + prefixLen, sizeof(formattedMsg) - prefixLen, msg, v);
-
-        // Invoke the Zephyr logging library to log the message.
-        //
-        // Unfortunately the Zephyr logging macros end up assigning uint16_t
-        // variables to uint16_t:10 fields, which triggers integer conversion
-        // warnings.  And treating the Zephyr headers as system headers does not
-        // help, apparently.  Just turn off that warning around this switch.
+    // Invoke the Zephyr logging library to log the message.
+    //
+    // Unfortunately the Zephyr logging macros end up assigning uint16_t
+    // variables to uint16_t:10 fields, which triggers integer conversion
+    // warnings.  And treating the Zephyr headers as system headers does not
+    // help, apparently.  Just turn off that warning around this switch.
 #pragma GCC diagnostic push
 #pragma GCC diagnostic ignored "-Wconversion"
-        switch (category)
-        {
-        case kLogCategory_Error:
-            LOG_ERR("%s", log_strdup(formattedMsg));
-            break;
-        case kLogCategory_Progress:
-        default:
-            LOG_INF("%s", log_strdup(formattedMsg));
-            break;
-        case kLogCategory_Detail:
-            LOG_DBG("%s", log_strdup(formattedMsg));
-            break;
-        }
-#pragma GCC diagnostic pop
+    switch (category)
+    {
+    case kLogCategory_Error:
+        LOG_ERR("%s", log_strdup(formattedMsg));
+        break;
+    case kLogCategory_Progress:
+    default:
+        LOG_INF("%s", log_strdup(formattedMsg));
+        break;
+    case kLogCategory_Detail:
+        LOG_DBG("%s", log_strdup(formattedMsg));
+        break;
     }
+#pragma GCC diagnostic pop
 
     // Let the application know that a log message has been emitted.
     DeviceLayer::OnLogOutput();
 }
 
+} // namespace Platform
 } // namespace Logging
 } // namespace chip
