+++ src/platform/logging/impl/android/Logging.cpp
@@ -1,0 +1,20 @@
+/* See Project chip LICENSE file for licensing information. */
+
+#include "platform/logging/LogV.h"
+#include "support/logging/Constants.h"
+
+#include <android/log.h>
+
+namespace chip {
+namespace Logging {
+namespace Platform {
+
+void LogV(const char * module, uint8_t category, const char * msg, va_list v)
+{
+    int priority = (category == kLogCategory_Error) ? ANDROID_LOG_ERROR : ANDROID_LOG_DEBUG;
+    __android_log_vprint(priority, module, msg, v);
+}
+
+} // namespace Platform
+} // namespace Logging
+} // namespace chip
