+++ src/qrcodetool/BUILD.gn
@@ -29,6 +29,7 @@ executable("qrcodetool") {
 
   public_deps = [
     "${chip_root}/src/lib/support",
+    "${chip_root}/src/platform/logging:stdio",
     "${chip_root}/src/setup_payload",
   ]
 }
