+++ src/transport/tests/TestSecureSession.cpp
@@ -183,13 +183,3 @@ int TestSecureSession()
 }
 
 CHIP_REGISTER_TEST_SUITE(TestSecureSession)
-
-namespace chip {
-namespace Logging {
-void __attribute__((weak)) LogV(uint8_t module, uint8_t category, const char * format, va_list argptr)
-{
-    (void) module, (void) category;
-    vfprintf(stderr, format, argptr);
-}
-} // namespace Logging
-} // namespace chip
